Fix inconsistency with robot rotation and movement.

+++ Headers/Obstacle.h
@@ -13,13 +13,13 @@ class Obstacle : public MoveableObject, public QGraphicsRectItem {
         qreal mp_coord_x;
         qreal mp_coord_y;
 
-        size_t mp_rotation;
+        qreal mp_rotation;
 
         QString mp_type;
 
         class PlayGround* mp_playground;
 
-        void do_rotation (const size_t angle);
+        void do_rotation (const qreal angle);
 
     public:
         Obstacle (const qreal width,
